requirements.txt
# Core Data Processing
pandas
numpy

# HTTP and API Libraries
requests
python-dotenv
groq
openai
google-cloud-aiplatform
vertexai

# Image Processing
Pillow

# Machine Learning and Analysis
sentence-transformers
scikit-learn
seaborn
matplotlib

# Progress Bars
tqdm

# File handling
pathlib

# Streamlit
streamlit